# docs/OpenAPI-Specification.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Family Tree API v1",
    "description": "المواصفة التفاعلية البرمجية القياسية لواجهات البرمجة في منصة شجرة العائلة الكبرى (Crowdsourced Family Tree Platform API).",
    "version": "1.0.0",
    "contact": {
      "name": "فريق تطوير شجرة العائلة",
      "email": "support@familytree.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "بيئة التطوير والاختبار المحلية"
    },
    {
      "url": "https://familytree.app/api/v1",
      "description": "بيئة الإنتاج الحية"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "تسجيل مستخدم جديد في المنظومة",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "full_name"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "user@example.com" },
                  "full_name": { "type": "string", "example": "عبدالله سلمان العلي" },
                  "phone": { "type": "string", "example": "+966500000000" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "تم تسجيل أو جلب المستخدم بنجاح" },
          "400": { "description": "بيانات الإدخال غير صالحة" }
        }
      }
    },
...
    "/claim/request": {
      "post": {
        "summary": "رفع طلب مطالبة وتوثيق ملكية بطاقة نسب",
        "responses": {
          "200": { "description": "تم تقديم طلب المطالبة بنجاح" },
          "409": { "description": "البطاقة مطالب بها مسبقاً" }
        }
      }
    },
...
  }
}